requirements.txt
matrix-nio==0.17
aioschedule==0.5.2
pillow==8.1.2
python-magic==0.4.22
requests==2.25.1